Handle 'no content' case.
Now that we use httr::http_type(), I believe it is impossible for content_type to be NULL or of zero length. The default content type reported by httr when none exists is "application/octet-stream". Checking for the length of the content seems good way to catch "no content".

Alternative possibility is to inspect the status and look for, e.g., "204 No Content".