Implement request_throttling()
Testing this revealed that I failed to set up `max_wait` correctly; it got discarded by the final line of `gh_build_request()`. And even if I had caught that, it still wasn't recorded in the response object so pagination could preserve the value.

So now the request captures the complete request list made up by gh and reuses that when paginating. I'll probably refactor this in the future to reuse the same httr2 request.